watchers/requirements.txt
# Bronze Tier
watchdog>=4.0.0

# Silver Tier — Gmail Watcher
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.0.0

# Silver Tier — WhatsApp + LinkedIn Watchers
playwright>=1.40.0
